+++ app/Enums/MediaCollectionType.php
@@ -5,4 +5,5 @@ namespace App\Enums;
 enum MediaCollectionType: string
 {
     case USER_PROFILE = 'user-profile';
+    case FILAMENT_MESSENGER = 'filament-messages';
 }