+++ app/Livewire/Traits/HasPollInterval.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Livewire\Traits;
+
+trait HasPollInterval
+{
+    public $pollInterval = '5s';
+
+    public function setPollInterval()
+    {
+        $this->pollInterval = config('filament-messages.poll_interval', '5s');
+    }
+}